Trait log4rs_rolling_file::roll::Roll

pub trait Roll: Debug + Send + Sync + 'static {
    fn roll(&self, file: &Path) -> Result<(), Box<Error>>;
}

A trait which processes log files after they have been rolled over.

Required Methods

fn roll(&self, file: &Path) -> Result<(), Box<Error>>

Processes the log file.

At the time that this method has been called, the log file has already been closed.

If this method returns successfully, there must no longer be a file at the specified location.